requirements.txt
boto3==1.4.7
botocore==1.7.20
docutils==0.14
python-dateutil==2.6.1
python-dotenv==0.7.1
s3transfer==0.1.11
